Moved show_help, show_info, show_about from popup module to main module.
